requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.10.1
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pytest==8.4.2
httpx==0.28.1